+++ src/rust/ide/view/graph-editor/src/component/node/input.rs
@@ -1,1 +1,4 @@
 pub mod area;
+
+pub use area::Expression;
+pub use area::Manager;
